Adds group assignment for LDAP/AD users.
The problem
-----------------
As of today, it is possible to configure default groups assignment for users in LDAP or AD directories. These groups are added when a user first logs in to JIRA, but with the authenticator, we skip JIRA authentication to rely on Shibboleth and the groups are never assigned.

The fix
---------
This feature enables this assignment to take place once per user by:
 * checking if a user has already been recorded as logged in JIRA
 * calling the update method related to the directoy if it is of the correct type of directory
 * recording a successful login for the user so as to update the user only once.